Fixed troubles with read/write. Added load test. Removed TreeNode
